refactor: reorganize server monitoring functionality and remove unused files
