+++ src/index.ts
@@ -1,12 +1,13 @@
 export { configManager } from "./config/config";
 export { UserConfig } from "./types/types";
 import { configManager } from "./config/config";
-import { startServerMonitoring } from "./core/status";
+import { startServerMonitoring } from "./core/status/status";
+import { logError } from "./utils/logger";
 
 const initializeScopeo = () => {
     const config= configManager.getConfig();
     if (!config) {
-        console.error("Scopeo config is not set");
+        logError("Scopeo config is not set");
         throw new Error("Scopeo config is not set");
     }
     startServerMonitoring();
